requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
cloudinary==1.32.0
cryptography==38.0.4
dj-database-url==1.2.0
Django==4.1.4
django-appconf==1.0.5
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-staticfiles==1.2.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
whitenoise==6.2.0
